% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permLR.R
\name{createPermGS}
\alias{createPermGS}
\title{createPermGS}
\usage{
createPermGS(B = 1000, restricted = TRUE, IP.method = "IPZ",
  imputeData = NULL, permuteData = NULL)
}
\arguments{
\item{B}{number of random permutations}

\item{restricted}{if TRUE only permute within strata}

\item{IP.method}{imputation/permuation method IPZ, IPT or none (default: IPZ)}

\item{imputeData}{user-supplied imputation function (ignored if IP.method is given)}

\item{permuteData}{user-supplied permutation function (ignore if IP.method is given)}
}
\value{
object of class permGS
}
\description{
Create permGS object representing a permutational group-sequential trial.
}
\examples{
## standard permutation test (no imputation, free permutations)
x <- createPermGS(1000, FALSE, "none")
## imputation using IPT method, restricted permutations
y <- createPermGS(1000, TRUE, "IPT")
}

