% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permLR.R
\name{impute.IPT}
\alias{impute.IPT}
\title{impute.IPT}
\usage{
impute.IPT(data)
}
\arguments{
\item{data}{matrix as returned by as.matrix(generateData(param))}
}
\value{
matrix containing imputed survival and censoring times (columns 1 and 2), and original treatment indicator (column 3)
}
\description{
Impute data according to IPT method. Output is supposed to be passed to permute.IPT
}

