% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPZ.R
\name{impute.IPZ}
\alias{impute.IPZ}
\title{impute.IPZ}
\usage{
impute.IPZ(data)
}
\arguments{
\item{data}{matrix as returned by as.matrix(generateData(param))}
}
\value{
original data with 4 new columns (V1 and V2) containing the imputed observations
}
\description{
Impute data according to IPZ method. Output is supposed to be passed to permute.IPZ
}
\references{
Wang, R., Lagakos, S.~W. and Gray, R.~J. Testing and interval estimation for two-sample survival comparisons with small sample sizes and unequal censoring. Biostatistics, 11(4), 676--692, January 2010.
}

