% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinze.R
\name{imputeHeinze}
\alias{imputeHeinze}
\title{imputeHeinze}
\usage{
imputeHeinze(data, pool = TRUE)
}
\arguments{
\item{data}{matrix as returned by as.matrix(generateData(param))}

\item{pool}{if TRUE impute events times from pooled Kaplan-Meier estimator (default: TRUE)}
}
\value{
list containing Kaplan-Meier estimators of censoring and survival distributions and the original data
}
\description{
Impute data according to Heinze et al. method. Output is supposed to be passed to permute.heinze
}
\references{
Heinze, G., Gnant, M. and Schemper, M. Exact Log-Rank Tests for Unequal Follow-Up. Biometrics, 59(4), December 2003.
}
