% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logrank_trafo}
\alias{logrank_trafo}
\title{Calculate ranks from right-censored survival data}
\usage{
logrank_trafo(time, event)
}
\arguments{
\item{time}{vector of survival times}

\item{event}{vector of event indicators (0=censored, 1=event)}
}
\value{
vector of ranks
}
\description{
Reduced version of coin::logrank_trafo
}

