% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parseFormula}
\alias{parseFormula}
\title{Parse formula of survival model}
\usage{
parseFormula(formula, data = parent.frame())
}
\arguments{
\item{formula}{formula object}

\item{data}{data.frame (optional)}
}
\value{
data.frame containing the parsed variables
}
\description{
Parse formula of survival model
}
