% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permLR.R
\name{permute.IPT}
\alias{permute.IPT}
\title{permute.IPT}
\usage{
permute.IPT(data, pp, index = TRUE)
}
\arguments{
\item{data}{matrix as returned by impute.IPT}

\item{pp}{vector of permuted indices}

\item{index}{not used}
}
\value{
matrix with time, status, trt columns
}
\description{
Permute survival times after imputation (IPT)
}

