% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permLR.R
\name{permute.IPZ}
\alias{permute.IPZ}
\title{permute.IPZ}
\usage{
permute.IPZ(data, pZ, index = FALSE)
}
\arguments{
\item{data}{matrix as returned by impute.IPT}

\item{pZ}{vector of permuted indices if index is TRUE, else binary vector of treatment assignments}

\item{index}{indicates if pZ is a vector of indices or a binary vector of treatment assignments}
}
\value{
matrix with time, status, Z columns
}
\description{
Permute treatment assignment after imputation (IPZ)
}

