% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinze.R
\name{permuteHeinze}
\alias{permuteHeinze}
\title{permuteHeinze}
\usage{
permuteHeinze(imp, pp, index = TRUE)
}
\arguments{
\item{imp}{list as returned by impute.heinze}

\item{pp}{vector of permuted indices}

\item{index}{not used}
}
\value{
matrix with time, status, trt columns
}
\description{
Perform single imputation and permutation step
}
\references{
Heinze, G., Gnant, M. and Schemper, M. Exact Log-Rank Tests for Unequal Follow-Up. Biometrics, 59(4), December 2003.
}
