% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPT.R
\name{permuteIPT}
\alias{permuteIPT}
\title{permuteIPT}
\usage{
permuteIPT(data, pp, index = TRUE)
}
\arguments{
\item{data}{matrix as returned by impute.IPT}

\item{pp}{vector of permuted indices}

\item{index}{not used}
}
\value{
matrix with time, status, trt columns
}
\description{
Permute survival times after imputation (IPT)
}
\references{
Wang, R., Lagakos, S.~W. and Gray, R.~J. Testing and interval estimation for two-sample survival comparisons with small sample sizes and unequal censoring. Biostatistics, 11(4), 676--692, January 2010.
}
