% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPZ.R
\name{permuteIPZ}
\alias{permuteIPZ}
\title{permuteIPZ}
\usage{
permuteIPZ(data, pZ, index = FALSE)
}
\arguments{
\item{data}{matrix as returned by impute.IPT}

\item{pZ}{vector of permuted indices if index is TRUE, else binary vector of treatment assignments}

\item{index}{indicates if pZ is a vector of indices or a binary vector of treatment assignments}
}
\value{
matrix with time, status, Z columns
}
\description{
Permute treatment assignment after imputation (IPZ)
}
\references{
Wang, R., Lagakos, S.~W. and Gray, R.~J. Testing and interval estimation for two-sample survival comparisons with small sample sizes and unequal censoring. Biostatistics, 11(4), 676--692, January 2010.
}
