% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sampleFromKM}
\alias{sampleFromKM}
\title{sampleFromKM}
\usage{
sampleFromKM(n, fit, start = 0, tmax = NULL, dv = 1)
}
\arguments{
\item{n}{sample size}

\item{fit}{Kaplan-Meier fit as returned by survfit}

\item{start}{if 0 sample from L(T), else sample from L(T, T > start)}

\item{tmax}{largest observation in pooled sample}

\item{dv}{1 if imputing events, 0 if imputing censoring times}
}
\value{
Random sample of survival times
}
\description{
Sample from distribution estimated by Kaplan-Meier estimator. Imputed values > tmax are right-censored.
}
