% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permLR.R
\name{shuffleBlock}
\alias{shuffleBlock}
\title{shuffleBlock
Permute block preserving group sizes, randomization blocks}
\usage{
shuffleBlock(block, strata = 0)
}
\arguments{
\item{block}{vector of row indices to be permuted}

\item{strata}{factor defining strata with block}
}
\value{
random permutation of each stratum within block
}
\description{
shuffleBlock
Permute block preserving group sizes, randomization blocks
}
