% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permLR.R
\name{summary.permGS}
\alias{summary.permGS}
\title{summary of permGS object}
\usage{
\method{summary}{permGS}(object, ...)
}
\arguments{
\item{object}{permGS object as returned by \code{\link{createPermGS}}}

\item{...}{additional parameters (currently unused)}
}
\value{
nothing
}
\description{
summary of permGS object
}
